refactor: Checkbox 제어 상태를 useControllableState 훅으로 통일
