+++ packages/jds/src/hooks/index.ts
@@ -1,2 +1,3 @@
+export { useControllableState } from "./useControllableState";
 export { useMediaQueryFlags } from "./useMediaQuery";
 export type { DeviceType, Breakpoints } from "./useMediaQuery";
